Separate posthog provider component
